# docs/openapi/scforge-public.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "SCForge \u5F00\u653E API",
    "description": "SCForge\uFF08\u751F\u5B58\u6218\u4E89\u63D2\u4EF6 / \u6A21\u7EC4\u8D44\u6E90\u5E73\u53F0\uFF09\u7684\u5BF9\u5916\u5F00\u653E\u63A5\u53E3\u3002\n\n**\u9274\u6743**\uFF1A\u8BFB\u63A5\u53E3\u533F\u540D\u53EF\u7528\uFF1B\u4E0E\u300C\u6211\u7684\u8D44\u6E90\u300D\u76F8\u5173\u7684\u8BFB\u5199\u9700\u8981 API Key \u2014\u2014 \u8BF7\u6C42\u5934 Authorization: Bearer scf_\u003Cbase64url\u003E\u3002\u7B7E\u53D1\u4E0E\u4F5C\u7528\u57DF\u89C1\u63A5\u5165\u6307\u5357\u3002\n\n**\u9519\u8BEF\u5F62\u72B6**\uFF1A\u6240\u6709\u9519\u8BEF\u54CD\u5E94\u7EDF\u4E00\u4E3A { \u0022detail\u0022: \u0022\u4E2D\u6587\u8BF4\u660E\u0022 }\uFF08\u89C1 ApiError.cs\uFF09\uFF0CHTTP \u72B6\u6001\u7801\u5373\u8BED\u4E49\uFF08400 \u53C2\u6570\u6216\u89C4\u5219\u4E0D\u5408\u6CD5\u3001401 \u672A\u767B\u5F55\u3001403 \u4F5C\u7528\u57DF\u4E0D\u8DB3\u3001404 \u8D44\u6E90\u4E0D\u5B58\u5728\u3001409 \u72B6\u6001\u51B2\u7A81\uFF09\u3002\n\n**\u65F6\u95F4**\uFF1A\u6240\u6709\u65F6\u95F4\u5B57\u6BB5\u6309\u5317\u4EAC\u65F6\u95F4\uFF08UTC\u002B8\uFF0C\u002B08:00\uFF09\u8F93\u51FA\u3002",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.cldery.com",
      "description": "\u751F\u4EA7\u73AF\u5883"
    }
  ],
  "paths": {
...
    "/scforge/addons/{addonId}/comments": {
      "get": {
        "tags": [
          "ScforgeComments"
        ],
        "summary": "\u8D44\u6E90\u7684\u8BC4\u8BBA\u6811\uFF08\u9876\u5C42\u8BC4\u8BBA \u002B \u5D4C\u5957\u56DE\u590D\uFF09\u3002",
        "parameters": [
          {
            "name": "addonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "ScforgeComments"
        ],
        "summary": "\u53D1\u8868\u8BC4\u8BBA\uFF1B\u5E26 parentId \u65F6\u662F\u5BF9\u67D0\u6761\u8BC4\u8BBA\u7684\u56DE\u590D\u3002",
        "parameters": [
          {
            "name": "addonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScforgeCommentCreateIn"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ScforgeCommentCreateIn"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/ScforgeCommentCreateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/scforge/comments/{id}": {
      "patch": {
        "tags": [
          "ScforgeComments"
        ],
        "summary": "\u7F16\u8F91\u81EA\u5DF1\u7684\u8BC4\u8BBA\u3002",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScforgeCommentUpdateIn"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ScforgeCommentUpdateIn"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/ScforgeCommentUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "ScforgeComments"
        ],
        "summary": "\u5220\u9664\u8BC4\u8BBA\u53CA\u5176\u5168\u90E8\u56DE\u590D\u3002",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/scforge/comments/{id}/vote/up": {
      "put": {
        "tags": [
          "ScforgeComments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/scforge/comments/{id}/vote/down": {
      "put": {
        "tags": [
          "ScforgeComments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/scforge/comments/{id}/vote": {
      "delete": {
        "tags": [
          "ScforgeComments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ScforgeCommentCreateIn": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "nullable": true
          },
          "parentId": {
            "type": "string",
            "description": "\u7236\u8BC4\u8BBA Id\uFF1A\u4E3A\u7A7A\u5373\u9876\u5C42\u8BC4\u8BBA\u3002",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "\u53D1\u8868\u8BC4\u8BBA\u6216\u56DE\u590D\u3002"
      },
      "ScforgeCommentUpdateIn": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "\u7F16\u8F91\u81EA\u5DF1\u8BC4\u8BBA\u7684\u6B63\u6587\u3002"
      },
...
    },
    "securitySchemes": {
      "ScforgeApiKey": {
        "type": "http",
        "description": "SCForge API Key\uFF08scf_ \u5F00\u5934\uFF09\u3002\u8BFB\u63A5\u53E3\u533F\u540D\u4E5F\u53EF\u8C03\u7528\uFF0C\u5E26\u4E0A Key \u624D\u80FD\u8BBF\u95EE\u300C\u6211\u7684\u8D44\u6E90\u300D\u4E0E\u5199\u63A5\u53E3\u3002",
        "scheme": "bearer",
        "bearerFormat": "scf_\u003Cbase64url\u003E"
      }
    }
  },
  "security": [
    {
      "ScforgeApiKey": []
    }
  ],
...
}